vicky: move get_poisoned_locks from Task to LockDatabase trait
