+++ vicky/src/lib/database/entities/lock.rs
@@ -99,4 +99,23 @@ pub mod db_impl {
             }
         }
     }
+
+    pub trait LockDatabase {
+        fn get_poisoned_locks(&mut self) -> Result<Vec<Lock>, VickyError>;
+        fn get_active_locks(&mut self) -> Result<Vec<Lock>, VickyError>;
+    }
+
+    impl LockDatabase for PgConnection {
+        fn get_poisoned_locks(&mut self) -> Result<Vec<Lock>, VickyError> {
+            use self::locks::dsl::*;
+
+            let poisoned_locks = {
+                let poisoned_db_locks: Vec<DbLock> =
+                    locks.filter(poisoned_by_task.is_not_null()).load(self)?;
+                poisoned_db_locks.into_iter().map(Lock::from).collect()
+            };
+
+            Ok(poisoned_locks)
+        }
+    }
 }
